+++ .changeset/mighty-mangos-exercise.md
@@ -1,0 +1,5 @@
+---
+'@bnb-chain/greenfield-js-sdk': patch
+---
+
+feat: Migrate bucket
