+++ packages/chain-sdk/src/clients/spclient/auth.ts
@@ -141,7 +141,7 @@ const SUPPORTED_HEADERS = [
   // HTTPHeaderAppDomain.toLocaleLowerCase(),
 ];
 
-// https://github.com/ethers-io/ethers.js/issues/823
+// https://github.com/ethers-io/ethers.js/discussions/4339
 export const secpSign = (digestBz: Uint8Array, privateKey: string) => {
   const signingKey = new SigningKey(privateKey);
   const signature = signingKey.signDigest(digestBz);
