+++ packages/chain-sdk/src/clients/spclient/spApis/migrateApproval.ts
@@ -1,0 +1,32 @@
+import { EMPTY_STRING_SHA256, METHOD_GET } from '@/constants';
+import { ICreateBucketMsgType, IMigrateBucketMsgType, ReqMeta } from '@/types';
+import { toHex, utf8ToBytes } from 'ethereum-cryptography/utils';
+
+// https://docs.bnbchain.org/greenfield-docs/docs/api/storgae-provider-rest/get_approval
+export const getMigrateMetaInfo = async (endpoint: string, msg: IMigrateBucketMsgType) => {
+  const path = '/greenfield/admin/v1/get-approval';
+  const query = 'action=MigrateBucket';
+  const url = `${endpoint}${path}?${query}`;
+  const unSignedMessageInHex = toHex(utf8ToBytes(JSON.stringify(msg)));
+
+  const reqMeta: Partial<ReqMeta> = {
+    contentSHA256: EMPTY_STRING_SHA256,
+    method: METHOD_GET,
+    url: {
+      hostname: new URL(url).hostname,
+      query,
+      path,
+    },
+    unsignMsg: unSignedMessageInHex,
+  };
+
+  const optionsWithOutHeaders: Omit<RequestInit, 'headers'> = {
+    method: METHOD_GET,
+  };
+
+  return {
+    url,
+    optionsWithOutHeaders,
+    reqMeta,
+  };
+};
